Move sample data download into separate script
in preparation of removing some of the local deployment complexities
